Fixed exercise 4.3.12.
The previous proof does not work in the case where y is much closer to another element of F, b, than x is to a, i.e, if g(y) - g(x) << 0. 
This is easily fixed by just picking as a comparison point the one for which g is smallest.